requirements.txt
anyio==4.4.0
asyncio==3.4.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
pyaes==1.6.1
pymongo==4.8.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
TgCrypto==1.2.5
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
